test permissions for namespaced controllers
